requirements.txt
pandas
datasets
openai
python-dotenv
tqdm
timeout-decorator
fuzzysearch
SQLAlchemy
regex
sentence-transformers
tabulate
torch
openpyxl
